python/requirements.txt
slack-sdk[socket-mode]>=3.27.0
flask>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.0
cachetools>=5.3.0
sentry-sdk>=2.0.0
pytest>=8.0.0
